+++ src/jsMain/kotlin/dev/inmo/jsuikit/modifiers/UIKitPlaceholder.kt
@@ -1,0 +1,5 @@
+package dev.inmo.jsuikit.modifiers
+
+object UIKitPlaceholder : UIKitModifier {
+    override val classes: Array<String> = arrayOf("uk-placeholder")
+}